% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnc.r
\name{cdnc}
\alias{cdnc}
\title{Cloud droplet number concentration (from Grosvenor et al, 2015).
Default values for many constants are taken from Brenguier et al (2000), leading to consistency with Eq. 1 from Quaas et al (2006).}
\usage{
cdnc(cod, re, k = 0.8, fad = 1, cw = 1.9e-06, Qext = 2)
}
\arguments{
\item{cod}{numeric, cloud optical depth}

\item{re}{numeric, cloud effective radius (in m)}

\item{k}{numeric, droplet size distribution shape parameter. Default is 0.8.}

\item{fad}{numeric, adiabaticity factor. Default is 1.}

\item{cw}{numeric, condensation rate (in kg m^-3 m^-3). Default is 1.9E-6.}

\item{Qext}{numeric, extinction efficiency factor. Default is 2.}
}
\value{
numeric, cloud droplet number concentration (in # m^-3)
}
\description{
Cloud droplet number concentration (from Grosvenor et al, 2015).
Default values for many constants are taken from Brenguier et al (2000), leading to consistency with Eq. 1 from Quaas et al (2006).
}
