% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cldphys.r
\docType{package}
\name{cldphys-package}
\alias{cldphys}
\alias{cldphys-package}
\title{Some helpful functions for cloud and atmospheric physics}
\description{
Here we just load dplyr so that the \%>\% operator is recognized throughout
}
\author{
\strong{Maintainer}: Odran Sourdeval \email{odran.sourdeval@univ-lille.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
