% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiab_condens_rate.r
\name{cw_adiab_bennartz}
\alias{cw_adiab_bennartz}
\title{Adiabatic condensation rate, following Bennartz et al (2009)}
\usage{
cw_adiab_bennartz(T, p, use_elr = FALSE)
}
\arguments{
\item{T}{numeric, air temperature (in K)}

\item{p}{numeric, atmospheric pressure (in Pa)}

\item{use_elr}{boolean, true if the environmental lapse rate (-6.5 K / km) should be used. If false, a moist adiabatic lapse rate will be used.}
}
\value{
numeric, adiabatic condentation rate in kg m-3 m-1 (kg )
}
\description{
Adiabatic condensation rate, following Bennartz et al (2009)
}
