% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.r
\name{density_ice}
\alias{density_ice}
\title{Density of ice
Parameterization from Pruppacher and Klett (1997)}
\usage{
density_ice(ta)
}
\arguments{
\item{t}{Temperature (in K)}
}
\value{
Density of ice (in kg/m^3)
}
\description{
Density of ice
Parameterization from Pruppacher and Klett (1997)
}
