% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.r
\name{density_wat}
\alias{density_wat}
\title{Density of water
Parameterization from Pruppacher and Klett (1997)}
\usage{
density_wat(ta)
}
\arguments{
\item{T}{Temperature (in K)}
}
\value{
Density of water (in kg/m^3)
}
\description{
Density of water
Parameterization from Pruppacher and Klett (1997)
}
