% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_vap_press.r
\name{es_i_koop}
\alias{es_i_koop}
\title{Saturation vapour pressure with respect to ice
Parameterization from Murphy and Koop (2005), Eq. 7
Considered valid for T > 110 K}
\usage{
es_i_koop(ta)
}
\arguments{
\item{ta}{numeric, air temperature (in K)}
}
\value{
numeric, saturation vapour pressure (in Pa)
}
\description{
Saturation vapour pressure with respect to ice
Parameterization from Murphy and Koop (2005), Eq. 7
Considered valid for T > 110 K
}
