% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_vap_press.r
\name{es_w_magnus}
\alias{es_w_magnus}
\title{Saturation vapour pressure with respect to water, using the Magnus formula}
\usage{
es_w_magnus(T)
}
\arguments{
\item{T}{numeric, air temperature (in K)}
}
\value{
numeric, saturation vapour pressure (in Pa)
}
\description{
Saturation vapour pressure with respect to water, using the Magnus formula
}
