% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{gibbs_diff_water_ice}
\alias{gibbs_diff_water_ice}
\title{Gibbs free energy for the transition from water to ice}
\usage{
gibbs_diff_water_ice(ta)
}
\arguments{
\item{ta}{air temperature (in K)}
}
\value{
Gibbs free energy (in J)
}
\description{
Gibbs free energy for the transition from water to ice
}
