% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{gibbs_vapor_ice}
\alias{gibbs_vapor_ice}
\title{Gibbs free energy for the transition from vapor to ice}
\usage{
gibbs_vapor_ice(r, ta, Si)
}
\arguments{
\item{r}{radius of the spherical ice embryo (in m)}

\item{ta}{air temperature (in K)}

\item{Si}{Supersaturation ratio with respect to ice}
}
\value{
Gibbs free energy (in J)
}
\description{
Gibbs free energy for the transition from vapor to ice
}
