% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_wat_nucleation.r
\name{gibbs_vapor_wat}
\alias{gibbs_vapor_wat}
\title{Gibbs free energy for the transition from vapor to water}
\usage{
gibbs_vapor_wat(r, ta, Sw)
}
\arguments{
\item{r}{radius of the spherical water embryo (in m)}

\item{ta}{air temperature (in K)}

\item{Sw}{Supersaturation ratio with respect to water}
}
\value{
Gibbs free energy (in J)
}
\description{
Gibbs free energy for the transition from vapor to water
}
