% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{gibbs_wat_ice}
\alias{gibbs_wat_ice}
\title{Gibbs free energy for the transition from water to ice}
\usage{
gibbs_wat_ice(r, ta, ratio_param = FALSE)
}
\arguments{
\item{r}{radius of the spherical ice embryo (in m)}

\item{ta}{air temperature (in K)}
}
\value{
Gibbs free energy (in J)
}
\description{
Gibbs free energy for the transition from water to ice
}
