% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_heat.r
\name{L_sub}
\alias{L_sub}
\title{Compute the sublimation latent heat
Parameters are from Murphy and Koop (2005)}
\usage{
L_sub(ta)
}
\arguments{
\item{ta}{numeric, air temperature (in K)}
}
\value{
numeric, latent heat of sublimation (in J / kg)
}
\description{
Compute the sublimation latent heat
Parameters are from Murphy and Koop (2005)
}
