% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_heat.r
\name{L_vap}
\alias{L_vap}
\title{Compute the vaporisation latent heat}
\usage{
L_vap(ta)
}
\arguments{
\item{ta}{numeric, air temperature (in K)}
}
\value{
numeric, latent heat of vaporisation (in J / kg)
}
\description{
Compute the vaporisation latent heat
}
