% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapse_rate.r
\name{lapse_rate_moist}
\alias{lapse_rate_moist}
\title{Moist adiabatic lapse rate (from Hemond and Fechner, 2015)}
\usage{
lapse_rate_moist(T, w)
}
\arguments{
\item{T}{numeric, air temperature (in K)}

\item{w}{numeric, mixing ratio (in kg kg^-1)}
}
\value{
numeric, moist adiabatic lapse rate in K m^-1
}
\description{
Moist adiabatic lapse rate (from Hemond and Fechner, 2015)
}
