% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_ratio.r
\name{mixing_ratio}
\alias{mixing_ratio}
\title{Water vapour mixing ratio}
\usage{
mixing_ratio(e, p)
}
\arguments{
\item{e}{numeric, partial pressure of water vapour (in Pa)}

\item{p}{numeric, atmospheric pressure (in Pa)}
}
\value{
numeric, water vapour mixing ratio (-)
}
\description{
Water vapour mixing ratio
}
