% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{n_ice}
\alias{n_ice}
\title{Number of water molecules forming the ice embryo}
\usage{
n_ice(r, ta)
}
\arguments{
\item{r}{radius of the spherical ice embryo (in m)}

\item{ta}{air temperature (in K)}
}
\value{
Number of water molecules forming the ice embryo
}
\description{
Number of water molecules forming the ice embryo
}
