% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_wat_nucleation.r
\name{n_wat}
\alias{n_wat}
\title{Number of water molecules forming the liquid water embryo}
\usage{
n_wat(r, ta)
}
\arguments{
\item{r}{radius of the spherical liquid water embryo (in m)}

\item{ta}{air temperature (in K)}
}
\value{
Number of water molecules forming the water embryo
}
\description{
Number of water molecules forming the liquid water embryo
}
