% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{nucl_rate_ice_hom}
\alias{nucl_rate_ice_hom}
\title{Nuleation rate for ice crystals from homogeneous nucleation
Theory by Pruppacher and Klett (1997), chapter 7}
\usage{
nucl_rate_ice_hom(Si, ta)
}
\arguments{
\item{Si}{Supersaturation ratio with respect to ice}

\item{ta}{air temperature (in K)}
}
\value{
Nucleation rate (in m^-3 s^-1)
}
\description{
Nuleation rate for ice crystals from homogeneous nucleation
Theory by Pruppacher and Klett (1997), chapter 7
}
