% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_wat_nucleation.r
\name{nucl_rate_wat_hom}
\alias{nucl_rate_wat_hom}
\title{Nuleation rate for water from homogeneous nucleation
Theory by Pruppacher and Klett (1997), chapter 7}
\usage{
nucl_rate_wat_hom(S, ta)
}
\arguments{
\item{S}{Supersaturation ratio with respect to water}

\item{ta}{air temperature (in K)}
}
\value{
Nucleation rate (in m^-3 s^-1)
}
\description{
Nuleation rate for water from homogeneous nucleation
Theory by Pruppacher and Klett (1997), chapter 7
}
