% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_wat_nucleation.r
\name{nucl_rate_wat_hom_miller}
\alias{nucl_rate_wat_hom_miller}
\title{Nuleation rate for water from homogeneous nucleation
Paramterization by Miller et al (1983), doi:10.1063/1.445236
From laboratory measurements, valid over the temperature range 230-290 K}
\usage{
nucl_rate_wat_hom_miller(S, ta)
}
\arguments{
\item{S}{Supersaturation ratio with respect to water}

\item{ta}{air temperature (in K)}
}
\value{
Nucleation rate (in m^-3 s^-1)
}
\description{
Nuleation rate for water from homogeneous nucleation
Paramterization by Miller et al (1983), doi:10.1063/1.445236
From laboratory measurements, valid over the temperature range 230-290 K
}
