% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{rc_vapor_ice}
\alias{rc_vapor_ice}
\title{Critical ice embryo radius for the transition from vapor to ice}
\usage{
rc_vapor_ice(ta, Si)
}
\arguments{
\item{ta}{air temperature (in K)}

\item{Si}{Supersaturation ratio with respect to ice}
}
\description{
Critical ice embryo radius for the transition from vapor to ice
}
