% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_wat_nucleation.r
\name{rc_vapor_wat}
\alias{rc_vapor_wat}
\title{Critical water embryo radius for the transition from vapor to liquid water}
\usage{
rc_vapor_wat(ta, Sw)
}
\arguments{
\item{ta}{air temperature (in K)}

\item{Sw}{Supersaturation ratio with respect to water}
}
\description{
Critical water embryo radius for the transition from vapor to liquid water
}
