% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{rc_wat_ice}
\alias{rc_wat_ice}
\title{Critical ice embryo radius for the transition from water to ice}
\usage{
rc_wat_ice(ta, ratio_param = FALSE)
}
\arguments{
\item{ta}{air temperature (in K)}
}
\description{
Critical ice embryo radius for the transition from water to ice
}
