% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.r
\name{spec_hum}
\alias{spec_hum}
\title{Specific humidity}
\usage{
spec_hum(e, p)
}
\arguments{
\item{e}{numeric, partial pressure of water vapour (in Pa)}

\item{p}{numeric, atmospheric pressure (in Pa)}
}
\value{
numeric, specitic humidity (kg kg^-1)
}
\description{
Specific humidity
}
