% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_tension.r
\name{surf_tens_ice_vap}
\alias{surf_tens_ice_vap}
\title{Surface tension of between ice and water vapor
Parameterization from Hale and Plummer (1974)}
\usage{
surf_tens_ice_vap(ta)
}
\arguments{
\item{ta}{Temperature (K)}
}
\value{
Surface tension (N/m)
}
\description{
Surface tension of between ice and water vapor
Parameterization from Hale and Plummer (1974)
}
