% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_tension.r
\name{surf_tens_ice_wat}
\alias{surf_tens_ice_wat}
\title{Surface tension of between ice and liquid water
Parameterization from Reinhardt and Doye (2013)}
\usage{
surf_tens_ice_wat(ta)
}
\arguments{
\item{ta}{Temperature (K)}
}
\value{
Surface tension (N/m)
}
\description{
Surface tension of between ice and liquid water
Parameterization from Reinhardt and Doye (2013)
}
