% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_tension.r
\name{surf_tens_wat_vap}
\alias{surf_tens_wat_vap}
\title{Surface tension of between liquid water and water vapor
Parameterization from Pruppacher and Klett (1997)}
\usage{
surf_tens_wat_vap(ta)
}
\arguments{
\item{ta}{Temperature (K)}
}
\value{
Surface tension (N/m)
}
\description{
Surface tension of between liquid water and water vapor
Parameterization from Pruppacher and Klett (1997)
}
