% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatures.r
\name{T_virtual}
\alias{T_virtual}
\title{Virtual temperature}
\usage{
T_virtual(T, qv)
}
\arguments{
\item{T}{numeric, air temperature (in K)}

\item{qv}{numeric, specific humidity (in kg kg^-1)}
}
\value{
numeric, virtual temperature (in K)
}
\description{
Virtual temperature
}
