% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_ice_nucleation.r
\name{v_ice}
\alias{v_ice}
\title{Volume of water molecules in ice}
\usage{
v_ice(ta)
}
\arguments{
\item{ta}{air temperature (in K)}
}
\value{
Volume of water molecules in ice (in m^3)
}
\description{
Volume of water molecules in ice
}
