% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_wat_nucleation.r
\name{v_wat}
\alias{v_wat}
\title{Volume of water molecules in liquid water}
\usage{
v_wat(ta)
}
\arguments{
\item{ta}{air temperature (in K)}
}
\value{
Volume of water molecules in water (in m^3)
}
\description{
Volume of water molecules in liquid water
}
